+Nasdaq functionality, db interfacing
NasdaqCommunicator parses and respects robots.txt.  Downloads and
parses the list of stock symbols. Controller then updates the relevant
table in the database.  NasdaqCommunicator retrieves the historical data
for a given symbol. The Controller then updates the database.

One shortcoming with the current setup: Program does not check for
updated historical information. Configuration is a one and done as-is.

RedditCommunicator has token retrieval down but is mostly incomplete.